Change display in InventoryActivity according to ShopActivity
