+++ app/src/main/java/ch/epfl/sweng/studyup/items/InventoryActivity.java
@@ -11,16 +11,15 @@ import android.widget.ArrayAdapter;
 import android.widget.ListView;
 
 import java.util.ArrayList;
-import java.util.List;
 
 import ch.epfl.sweng.studyup.R;
-import ch.epfl.sweng.studyup.player.Player;
-import ch.epfl.sweng.studyup.questions.DisplayQuestionActivity;
+import ch.epfl.sweng.studyup.utils.ListItemAdapter;
 import ch.epfl.sweng.studyup.utils.navigation.NavigationStudent;
 import static ch.epfl.sweng.studyup.utils.Constants.*;
 
 public class InventoryActivity extends NavigationStudent {
     private ArrayAdapter<String> adapter;
+    private ListItemAdapter listItemAdapter;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -34,14 +33,14 @@ public class InventoryActivity extends NavigationStudent {
         navigationSwitcher(InventoryActivity.this, InventoryActivity.class, INVENTORY_INDEX);
 
         ArrayList<String> itemsName = Items.getItemsNames();
-        ListView listView = findViewById(R.id.listViewItems);
-        adapter = new ArrayAdapter<>(this, android.R.layout.simple_list_item_1, itemsName);
-        listView.setAdapter(adapter);
-
-        listView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
+        ListView ownedItems = findViewById(R.id.listViewItems);
+        listItemAdapter = new ListItemAdapter(getApplicationContext(), Items.values(), false);
+        ownedItems.setAdapter(listItemAdapter);
+        ownedItems.setOnItemClickListener(new AdapterView.OnItemClickListener() {
             @Override
             public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
-                startActivity(new Intent(parent.getContext(), DisplayItemActivity.class).putExtra(DisplayItemActivity.class.getName(), adapter.getItem(position)));
+                Items item = (Items)listItemAdapter.getItem(position);
+                startActivity(new Intent(parent.getContext(), DisplayItemActivity.class).putExtra(DisplayItemActivity.class.getName(), item.getName()));
             }
         });
     }
