+++ app/src/main/java/ch/epfl/sweng/studyup/utils/ListItemAdapter.java
@@ -7,46 +7,56 @@ import android.widget.BaseAdapter;
 import android.widget.ImageView;
 import android.widget.TextView;
 
-import java.util.ArrayList;
-
 import ch.epfl.sweng.studyup.R;
+import ch.epfl.sweng.studyup.items.Items;
+import ch.epfl.sweng.studyup.player.Player;
 
 public class ListItemAdapter extends BaseAdapter {
-
     private Context cnx;
-    private ArrayList<String> names;
-    private ArrayList<Integer> ids;
+    private Items[] items;
+    private boolean displayPriceAndCoin;
 
-    public ListItemAdapter(Context cnx, ArrayList<String> names, ArrayList<Integer> ids) {
+    public ListItemAdapter(Context cnx, Items[] items, boolean displayPriceAndCoin) {
         this.cnx=cnx;
-        this.names=names;
-        this.ids=ids;
+        this.items = items;
+        this.displayPriceAndCoin = displayPriceAndCoin;
     }
 
     @Override
     public int getCount() {
-        return names.size();
+        return items.length;
     }
 
     @Override
     public Object getItem(int position) {
-        return names.get(position);
+        return items[position];
     }
 
     @Override
     public long getItemId(int position) {
-        return ids.get(position);
+        return 0;
     }
 
     @Override
     public View getView(int position, View convertView, ViewGroup parent) {
         if(convertView==null){
-            convertView=View.inflate(cnx, R.layout.quest_item_model, null);
+            convertView=View.inflate(cnx, R.layout.shop_list_item_model, null);
         }
-        ImageView image_view = (ImageView) convertView.findViewById(R.id.is_quest_done);
-        TextView text_view = (TextView) convertView.findViewById(R.id.quest_title);
-        image_view.setImageResource(ids.get(position));
-        text_view.setText(names.get(position));
+        Items item = items[position];
+        ImageView shopItemImage = convertView.findViewById(R.id.shop_item_image);
+        TextView shopItemName =  convertView.findViewById(R.id.shop_item_name);
+        shopItemImage.setImageResource(item.getImageName());
+        shopItemName.setText(item.getName());
+        if(displayPriceAndCoin) {
+            ImageView shopCoin = convertView.findViewById(R.id.shop_coin);
+            shopCoin.setImageResource(R.drawable.gold_coin);
+            TextView shopItemPrice =  convertView.findViewById(R.id.shop_item_price);
+            shopItemPrice.setText(Integer.toString(item.getPrice()));
+        } else {
+            TextView inventoryItemQuantity = convertView.findViewById(R.id.inventory_item_quantity);
+            String quantity = Integer.toString(Items.itemsToQuantity(Player.get().getItems()).get(item)); //getting the quantity
+            inventoryItemQuantity.setText(quantity + "x");
+        }
         return convertView;
     }
 }
