+++ app/src/main/res/layout/shop_list_item_model.xml
@@ -35,7 +35,7 @@
         android:layout_width="61dp"
         android:layout_height="68dp"
         android:text="TextView"
-        android:textSize="50sp"
+        android:textSize="40sp"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.163"
@@ -45,8 +45,8 @@
 
     <ImageView
         android:id="@+id/shop_coin"
-        android:layout_width="52dp"
-        android:layout_height="69dp"
+        android:layout_width="50dp"
+        android:layout_height="50dp"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.0"
@@ -54,5 +54,16 @@
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintVertical_bias="0.0"
         app:srcCompat="@drawable/gold_coin" />
+
+    <TextView
+        android:id="@+id/inventory_item_quantity"
+        android:layout_width="78dp"
+        android:layout_height="66dp"
+        android:text="TextView"
+        android:textSize="50dp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.81"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
 
 </android.support.constraint.ConstraintLayout>
